IKASAN-1809 fixing null pointer issue

+++ ikasaneip/visualisation/dashboard/src/main/java/org/ikasan/dashboard/ui/visualisation/component/BusinessStreamUploadDialog.java
@@ -143,7 +143,9 @@ public class BusinessStreamUploadDialog extends AbstractCloseableResizableDialog
                 return;
             }
 
-            this.businessStreamMetaDataService.delete(businessStreamMetaData.getId());
+            if(businessStreamMetaData != null) {
+                this.businessStreamMetaDataService.delete(businessStreamMetaData.getId());
+            }
 
             BusinessStreamMetaData saveBusinessStreamMetaData = new BusinessStreamMetaDataImpl();
             saveBusinessStreamMetaData.setId(businessStreamNameTextfield.getValue());
